增加基于Mysql表格的ID生成实现，及相应的测试用例

+++ src/main/java/net/techgy/idgenerator/IDAssignmentMapper.java
@@ -9,7 +9,7 @@ public interface IDAssignmentMapper {
 
 	Set<IIDAssignment> loadAllUsingAssigment();
 	
-    int maxPrefixValue(String clientId,String entitId);
+    long maxPrefixValue(String clientId,String entitId);
 	
 	boolean isExistPrefix(String clientId,String entitId);
